Extract ValidationRules interfaces It resolves circular d.ts dependency
